# car.hire.rental/client.core/src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: Car Hire API
  version: 0.1.0
  contact:
    name: Vlad Carcu
  description: Car Hire API
servers:
  - url: http://localhost:{port}
    description: local environment
    variables:
      port:
        default: '8080'
...
paths:
...
  /user/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32

    # Get a user by id
    get:
      tags:
        - user
      summary: get a user
      description: returns the details for a user, based on id, except the password
      operationId: getUser
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    # Delete a user by id
    delete:
      tags:
        - user
      summary: delete a user
      description: soft deletes a user
      operationId: deleteUser
      responses:
        '200':
          description: success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    # Add a new user
    patch:
      tags:
        - user
      summary: update an existing user
      description: allows updating an existing user. Existing fields can't be set to null. The id, username, password and status are always ignored.
      operationId: patchUser
      requestBody:
        description: The fields of the user to change
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
            example:
              email: vladcarcu@email.com
              firstName: Vlad
              lastName: Carcu
              countryOfResidence:
                id: 2
              driverLicense: BV123456
      responses:
        '200':
          description: success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  responses:
...
    NotFound:
      description: the results are either empty or couldn't be found
    Unauthorized:
      description: unauthorized
    BadRequest:
      description: bad request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ValidationDTO'
    ServerError:
      description: internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDTO'
  schemas:
    ValidationDTO:
      type: object
      properties:
        code:
          type: string
        field:
          type: string
        description:
          type: string
    ErrorDTO:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: user id
        email:
          type: string
          format: email
          example: vladcarcu@email.com
          description: user email address
        username:
          type: string
          example: vladcarcu
        password:
          type: string
          format: password
          example: secret
        firstName:
          type: string
          example: Vlad
          description: the first name of the user
        lastName:
          type: string
          example: Carcu
          description: the last name of the user
        countryOfResidence:
          $ref: '#/components/schemas/CountryDTO'
        driverLicense:
          type: string
          example: BV123456
          description: the serie and number of the user's driver's license
        status:
          $ref: '#/components/schemas/UserStatus'
    CountryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: country id
        name:
          type: string
          example: Romania
          description: the name of the country
        isoCode:
          type: string
          example: ROU
          description: the iso code of the country
    UserStatus:
      type: string
      enum:
        - ACTIVE
        - DELETED
...